+++ Sources/BraintreeThreeDSecure/Public/BraintreeThreeDSecure/BraintreeThreeDSecure.h
@@ -17,12 +17,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeThreeDSecureVersionString[];
 #import <Braintree/BTThreeDSecurePostalAddress.h>
 #import <Braintree/BTThreeDSecureRequest.h>
 #import <Braintree/BTThreeDSecureResult.h>
-#import <Braintree/BTThreeDSecureV2UICustomization.h>
-#import <Braintree/BTThreeDSecureV2ButtonCustomization.h>
-#import <Braintree/BTThreeDSecureV2LabelCustomization.h>
-#import <Braintree/BTThreeDSecureV2TextBoxCustomization.h>
-#import <Braintree/BTThreeDSecureV2ToolbarCustomization.h>
-#import <Braintree/BTThreeDSecureV2BaseCustomization.h>
 #else
 #import <BraintreePaymentFlow/BraintreePaymentFlow.h>
 #import <BraintreeCard/BraintreeCard.h>
@@ -32,10 +26,4 @@ FOUNDATION_EXPORT const unsigned char BraintreeThreeDSecureVersionString[];
 #import <BraintreeThreeDSecure/BTThreeDSecurePostalAddress.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureRequest.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureResult.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2ButtonCustomization.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2LabelCustomization.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2TextBoxCustomization.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2ToolbarCustomization.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2BaseCustomization.h>
 #endif