+++ UnitTests/BraintreeThreeDSecureTests/BraintreeThreeDSecureTests-Bridging-Header.h
@@ -5,8 +5,6 @@
 #import <BraintreeThreeDSecure/BTThreeDSecureAdditionalInformation_Internal.h>
 #import <BraintreeThreeDSecure/BTThreeDSecurePostalAddress_Internal.h>
 #import <BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure_Internal.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization_Internal.h>
-#import <BraintreeThreeDSecure/BTThreeDSecureV2BaseCustomization_Internal.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureRequest.h>
 #import <BraintreePaymentFlow/BTPaymentFlowClient.h>
 #import <BraintreeCore/BraintreeCore-Swift.h>