+++ UnitTests/BraintreeThreeDSecureTests/ThreeDSecureV2Customization_Tests/BTThreeDSecureV2TextBoxCustomization_Tests.swift
@@ -1,5 +1,6 @@
 import XCTest
 import CardinalMobile
+@testable import BraintreeThreeDSecure
 
 class BTThreeDSecureV2TextBoxCustomization_Tests: XCTestCase {